+++ timer.h
@@ -4,17 +4,18 @@
 
 class Timer {
     private:
-        int second;
-        int minute;
-        int hour;
+        int8_t second;
+        int8_t minute;
+        int8_t hour;
         unsigned long previous_millis;
         unsigned long current_millis;
         bool is_running;
+        
     public:
         Timer();
         ~Timer();
-        void curUp(int* pos_cursor);
-        void curDown(int* pos_cursor);
+        void curUp(int8_t* pos_cursor);
+        void curDown(int8_t* pos_cursor);
         void toggle();
         void secondCount();
         void print(LiquidCrystal* lcd) const;
